Add files via upload
This model detects what type of clothes is in the photo. Two methods were used Convolutional Neural Networks and Multilayer perceptron. The program compares the performance of each of them in terms of accuracy and time. It concludes that in this case, a neural network has a higher accuracy but also slowest time, while MLP comes with fast running time but a significant lower accuracy. Therefore, I can conclude that CNN is more suitable in this case.